Refactor role constant
