+++ src/main/java/id/ac/itb/logistik/ditlog/controller/ContractController.java
@@ -4,10 +4,8 @@ import id.ac.itb.logistik.ditlog.model.BaseResponse;
 import id.ac.itb.logistik.ditlog.model.SPMKContract;
 import id.ac.itb.logistik.ditlog.model.User;
 import id.ac.itb.logistik.ditlog.repository.SPMKContractRepository;
-import id.ac.itb.logistik.ditlog.service.RoleConstant;
+import id.ac.itb.logistik.ditlog.model.RoleConstant;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
